[CI Visibility] Harden CODEOWNERS path anchoring, lazy descendant regex, and move fallback tests to unit test project
